feat: create db seeder and split in feature modules
Ref #S202111OQM-12
