+++ backend/src/app.module.ts
@@ -2,14 +2,19 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { ServicesModule } from './features/services/services.module';
+import { DesksModule } from './features/desks/desks.module';
 
 @Module({
   imports: [
     TypeOrmModule.forRoot({
       type: 'sqlite',
-      database: 'queue.db',
+      database: 'db.sqlite',
+      autoLoadEntities: true,
       synchronize: true,
     }),
+    ServicesModule,
+    DesksModule,
   ],
   controllers: [AppController],
   providers: [AppService],
