+++ backend/src/features/desks/desk.entity.ts
@@ -1,5 +1,5 @@
 import { Column, Entity, ManyToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { Service } from '../service/service.entity';
+import { Service } from '../services/service.entity';
 
 @Entity()
 export class Desk {
@@ -9,7 +9,7 @@ export class Desk {
   /**
    * The services available at this Desk
    */
-  @ManyToMany((type) => Service, (service) => service.desks)
+  @ManyToMany(() => Service, service => service.desks)
   services: Service[];
 
   /**
