+++ backend/src/features/services/service.entity.ts
@@ -1,5 +1,11 @@
-import { Column, Entity, ManyToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { Desk } from '../desk/desk.entity';
+import {
+  Column,
+  Entity,
+  ManyToMany,
+  PrimaryGeneratedColumn,
+  JoinTable,
+} from 'typeorm';
+import { Desk } from '../desks/desk.entity';
 
 @Entity()
 export class Service {
@@ -28,6 +34,7 @@ export class Service {
   /**
    * The desks where this service is available
    */
-  @ManyToMany((type) => Desk, (desk) => desk.services)
+  @ManyToMany(() => Desk, desk => desk.services, { cascade: true, eager: true })
+  @JoinTable()
   desks: Desk[];
 }
