+++ backend/src/features/services/services.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Service } from './service.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Service])],
+})
+export class ServicesModule {}
